# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Yzu视频 Premium Video Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category (owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (owner)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category (owner)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ── SUBSCRIPTIONS ─────────────────────────────────────────────────────────
...
  /bundles:
    get:
      operationId: listBundles
      tags: [bundles]
      summary: List video bundles (purchase-only packs of 1-10 videos)
      responses:
        "200":
          description: Bundle list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bundle"
    post:
      operationId: createBundle
      tags: [bundles]
      summary: Create a bundle (owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BundleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleDetail"
  /bundles/{id}:
    get:
      operationId: getBundle
      tags: [bundles]
      summary: Get bundle detail including its videos
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bundle detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleDetail"
    patch:
      operationId: updateBundle
      tags: [bundles]
      summary: Update a bundle (owner)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BundleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleDetail"
    delete:
      operationId: deleteBundle
      tags: [bundles]
      summary: Delete a bundle (owner)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /bundles/{id}/purchase:
    post:
      operationId: purchaseBundle
      tags: [bundles]
      summary: Purchase a bundle using wallet balance (grants permanent access to its videos)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundlePurchaseResult"

  # ── WALLET ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    BundleVideoItem:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
        title:
          type: string
        thumbnail:
          type: ["string", "null"]
    Bundle:
      type: object
      required: [id, title, price, isActive, sortOrder, videoCount, discountPercent, hasPurchased, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        badge:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
        videoCount:
          type: integer
        discountPercent:
          type: integer
        hasPurchased:
          type: boolean
        createdAt:
          type: string
          format: date-time
    BundleDetail:
      type: object
      required: [id, title, price, isActive, sortOrder, videoCount, discountPercent, hasPurchased, videos, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        badge:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
        videoCount:
          type: integer
        discountPercent:
          type: integer
        hasPurchased:
          type: boolean
        videos:
          type: array
          items:
            $ref: "#/components/schemas/BundleVideoItem"
        createdAt:
          type: string
          format: date-time
    BundleInput:
      type: object
      required: [title, price, videoIds]
      properties:
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        badge:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
        videoIds:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: string
    BundleUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        badge:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
        videoIds:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: string
    BundlePurchaseResult:
      type: object
      required: [id, userId, bundleId, price, createdAt, bundle]
      properties:
        id:
          type: string
        userId:
          type: string
        bundleId:
          type: string
        price:
          type: number
        createdAt:
          type: string
          format: date-time
        bundle:
          $ref: "#/components/schemas/BundleDetail"
...
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        videoCount:
          type: integer
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    # ── SUBSCRIPTION ────────────────────────────────────────────────────────
...