# lib/api-zod/src/generated/types/bundlePurchaseResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Yzu视频 Premium Video Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BundleDetail } from './bundleDetail';

export interface BundlePurchaseResult {
  id: string;
  userId: string;
  bundleId: string;
  price: number;
  createdAt: Date;
  bundle: BundleDetail;
}