# lib/api-zod/src/generated/types/bundleUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Yzu视频 Premium Video Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BundleUpdate {
  title?: string;
  description?: string;
  thumbnail?: string;
  price?: number;
  originalPrice?: number;
  badge?: string;
  isActive?: boolean;
  sortOrder?: number;
  /**
     * @minItems 1
     * @maxItems 10
     */
  videoIds?: string[];
}